Core/Random: Changed random functions returning doubles to return floats
* They were all cast to float at use anyway
* Improves roll_chance_f performance (rand32() is now called internally by uniform_real_distribution once instead of twice)

+++ src/server/game/Pools/PoolMgr.cpp
@@ -272,7 +272,7 @@ void PoolGroup<T>::SpawnObject(ActivePoolData& spawns, uint32 limit, uint32 trig
     // Attempt to select one object based on explicit chance.
     if (!ExplicitlyChanced.empty())
     {
-        float roll = (float)rand_chance();
+        float roll = rand_chance();
         for (PoolObject& candidate : candidates)
         {
             if (candidate.chance > 0)
